+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ CHANGELOG
 
 Release NEXT
 ------------
+- Make external and internal ips a property. Force all models inherited from inherited VirtualMachineMixin to define it.
 
 Release 0.124.0
 ---------------